requirements.txt
bs4~=0.0.2
beautifulsoup4~=4.12.3
soupsieve~=2.6
h11~=0.14.0
numpy~=2.1.0
typing_extensions~=4.12.2
pillow~=10.4.0
pip~=23.2.1
distro~=1.9.0
nltk~=3.9.1
regex~=2024.7.24
scipy~=1.14.1
scikit-learn~=1.5.1
click~=8.1.7
tqdm~=4.66.5
requests~=2.32.3
pytz~=2024.1
colorama~=0.4.6
sniffio~=1.3.1
httpcore~=1.0.5
idna~=3.8
certifi~=2024.8.30
jiter~=0.5.0
threadpoolctl~=3.5.0
joblib~=1.4.2
openai~=1.45.1
httpx~=0.27.2
pydantic~=2.9.1
anyio~=4.4.0
pandas~=2.2.2
pydantic_core~=2.23.3
python-dateutil~=2.9.0.post0
urllib3~=2.2.2
six~=1.16.0
textblob~=0.18.0.post0
termcolor~=2.4.0
pytesseract~=0.3.13
packaging~=24.1